Time: 79 ms (34.30%), Space: 54.8 MB (59.62%) - LeetHub

+++ 0493-reverse-pairs/0493-reverse-pairs.java
@@ -39,7 +39,7 @@ class Solution {
         int right = mid + 1;
         int cnt = 0;
         for (int i = low; i <= mid; i++) {
-            while (right <= high && (long)arr[i] > 2L * arr[right]) right++;
+            while (right <= high && arr[i] > 2L * arr[right]) right++;
             cnt += (right - (mid + 1));
         }
         return cnt;
